+++ packages/components/atoms/f-button/test/accessibility/axe-accessibility.spec.js
@@ -2,31 +2,29 @@ import ActionButton from '../../test-utils/component-objects/f-button--action.co
 import LinkButton from '../../test-utils/component-objects/f-button--link.component';
 import IconButton from '../../test-utils/component-objects/f-button--icon.component';
 
-const { getAxeResults } = require('../../../../../../test/utils/axe-helper');
-
 describe('Accessibility tests', () => {
-    it('a11y - should test f-button action component WCAG compliance', () => {
+    it('a11y - should test f-button action component WCAG compliance', async () => {
         // Act
-        ActionButton.load();
-        const axeResults = getAxeResults('f-button - action');
+        await ActionButton.load();
+        const axeResults = await ActionButton.getAxeResults('f-button - action');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);
     });
 
-    it('a11y - should test f-button link component WCAG compliance', () => {
+    it('a11y - should test f-button link component WCAG compliance', async () => {
         // Act
-        LinkButton.load();
-        const axeResults = getAxeResults('f-button - link');
+        await LinkButton.load();
+        const axeResults = await LinkButton.getAxeResults('f-button - link');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);
     });
 
-    it('a11y - should test f-button icon component WCAG compliance', () => {
+    it('a11y - should test f-button icon component WCAG compliance', async () => {
         // Act
-        IconButton.load();
-        const axeResults = getAxeResults('f-button - icon');
+        await IconButton.load();
+        const axeResults = await IconButton.getAxeResults('f-button - icon');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);